We do a little ABOMINATION OF FILE MERGING
